Apps/GroundSoftware/requirements.txt
mypy
typing_extensions>=4.0.0
typeguard
ipykernel
scapy
python-pcapng
numba~=0.56.4
numpy<1.24
matplotlib
seaborn~=0.12.2
pandas
tabulate
styleframe
xlsxwriter
pynput
multi-await
plotly
dash
panel
holoviews
hvplot
datashader
lxml
pytest
pytest-asyncio
coverage
pytest-cov
requests
# Need Protobuf <=3.20.x (after that is 4, which breaks yamcs protos)
protobuf <= 3.20.999
yamcs-client==1.7.5
jsonpickle
termcolor
colorama
verboselogs
coloredlogs~=15.0.1
ulid-py
pyserial
pyserial-asyncio
ipykernel
bitstruct
pyzmq
attrs
argparse
streamz
pybgapi==1.2.0
tables
dill[graph]
pdoc3
opencv-python